requirements.txt
google-generativeai
python-dotenv
opencv-python
pyyaml
pillow
numpy